+++ cabal.project
@@ -13,6 +13,7 @@ packages: ./typed-protocols
           ./ouroboros-consensus-byron
           ./ouroboros-consensus-byronspec
           ./ouroboros-consensus-shelley
+          ./ouroboros-consensus-shelley-test
           ./ouroboros-consensus-cardano
           ./io-sim
           ./io-sim-classes
@@ -69,8 +70,10 @@ package ouroboros-consensus-byron
   flags: +asserts
 
 package ouroboros-consensus-shelley
-  tests: True
   flags: +asserts
+
+package ouroboros-consensus-shelley-test
+  tests: True
 
 package ouroboros-consensus-cardano
   flags: +asserts
