+++ ouroboros-consensus-shelley-test/ouroboros-consensus-shelley-test.cabal
@@ -1,0 +1,114 @@
+name:                  ouroboros-consensus-shelley-test
+version:               0.1.0.0
+synopsis:              Test infrastructure for Shelley
+-- description:
+license:               Apache-2.0
+license-files:
+  LICENSE
+  NOTICE
+copyright:             2020 Input Output (Hong Kong) Ltd.
+author:                IOHK Engineering Team
+maintainer:            operations@iohk.io
+category:              Network
+build-type:            Simple
+cabal-version:         >=1.10
+
+source-repository head
+  type:     git
+  location: https://github.com/input-output-hk/ouroboros-network
+
+library
+  hs-source-dirs:      src
+
+  exposed-modules:
+                       Test.Consensus.Shelley.Examples
+                       Test.Consensus.Shelley.Generators
+                       Test.Consensus.Shelley.MockCrypto
+                       Test.ThreadNet.Infra.Shelley
+                       Test.ThreadNet.TxGen.Shelley
+
+  build-depends:       base              >=4.9   && <4.13
+                     , binary            >=0.8   && <0.9
+                     , bytestring        >=0.10  && <0.11
+                     , cardano-binary
+                     , cardano-crypto-class
+                     , cardano-crypto-wrapper
+                     , cardano-prelude
+                     , cardano-slotting
+                     , containers        >=0.5   && <0.7
+                     , cryptonite        >=0.25  && <0.26
+                     , generic-random
+                     , hedgehog-quickcheck
+                     , iproute
+                     , mtl               >=2.2   && <2.3
+                     , QuickCheck
+                     , time
+                     , serialise         >=0.2   && <0.3
+                     , transformers
+
+                       -- cardano-ledger-specs
+                     , shelley-spec-ledger
+                     , shelley-spec-ledger-test
+
+                     , ouroboros-network
+                     , ouroboros-consensus
+                     , ouroboros-consensus-test-infra
+                     , ouroboros-consensus-shelley
+
+  default-language:    Haskell2010
+  ghc-options:         -Wall
+                       -Wcompat
+                       -Wincomplete-uni-patterns
+                       -Wincomplete-record-updates
+                       -Wpartial-fields
+                       -Widentities
+                       -Wredundant-constraints
+                       -Wmissing-export-lists
+                       -fno-ignore-asserts
+
+test-suite test
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      test
+  main-is:             Main.hs
+  other-modules:
+                       Test.Consensus.Shelley.Golden
+                       Test.Consensus.Shelley.Serialisation
+                       Test.ThreadNet.RealTPraos
+
+  build-depends:       base
+                     , binary
+                     , bytestring
+                     , cardano-binary
+                     , cardano-crypto-wrapper
+                     , cardano-prelude
+                     , cardano-slotting
+                     , cborg             >=0.2.2 && <0.3
+                     , containers
+                     , QuickCheck
+                     , time
+                     , tasty
+                     , tasty-hunit
+                     , tasty-quickcheck
+
+                       -- cardano-ledger-specs
+                     , shelley-spec-ledger
+                     , shelley-spec-ledger-test
+
+                     , ouroboros-network
+                     , ouroboros-consensus
+                     , ouroboros-consensus-test-infra
+                     , ouroboros-consensus-shelley
+                     , ouroboros-consensus-shelley-test
+
+  default-language:    Haskell2010
+  ghc-options:         -Wall
+                       -Wcompat
+                       -Wincomplete-uni-patterns
+                       -Wincomplete-record-updates
+                       -Wpartial-fields
+                       -Widentities
+                       -Wredundant-constraints
+                       -Wmissing-export-lists
+                       -fno-ignore-asserts
+                       -threaded
+                       -rtsopts
