+++ ouroboros-consensus-shelley-test/src/Test/Consensus/Shelley/Generators.hs
@@ -1,21 +1,18 @@
-{-# LANGUAGE DisambiguateRecordFields #-}
-{-# LANGUAGE FlexibleInstances        #-}
-{-# LANGUAGE GADTs                    #-}
-{-# LANGUAGE LambdaCase               #-}
-{-# LANGUAGE NamedFieldPuns           #-}
-{-# LANGUAGE OverloadedStrings        #-}
-{-# LANGUAGE PatternSynonyms          #-}
-{-# LANGUAGE ScopedTypeVariables      #-}
-{-# LANGUAGE StandaloneDeriving       #-}
-{-# LANGUAGE TypeApplications         #-}
+{-# LANGUAGE FlexibleInstances   #-}
+{-# LANGUAGE GADTs               #-}
+{-# LANGUAGE NamedFieldPuns      #-}
+{-# LANGUAGE OverloadedStrings   #-}
+{-# LANGUAGE PatternSynonyms     #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE StandaloneDeriving  #-}
+{-# LANGUAGE TypeApplications    #-}
 
 {-# OPTIONS_GHC -Wno-orphans #-}
-module Test.Consensus.Shelley.Ledger (tests) where
+module Test.Consensus.Shelley.Generators (
+    SomeResult (..)
+  , genHash
+  ) where
 
-import qualified Codec.CBOR.Write as CBOR
-import qualified Data.Binary.Get as Get
-import qualified Data.Binary.Put as Put
-import qualified Data.ByteString.Lazy as Lazy
 import           Data.IP (IPv4, IPv6, toIPv4, toIPv6)
 import qualified Data.Map.Strict as Map
 import           Data.Maybe (fromJust)
@@ -23,22 +20,18 @@ import           Data.Proxy (Proxy (..))
 import           Data.Ratio ((%))
 import           Data.Sequence.Strict (StrictSeq)
 import qualified Data.Sequence.Strict as StrictSeq
-import           Data.Word (Word64, Word8)
+import           Data.Word (Word8)
 import           Numeric.Natural (Natural)
 
-import           Cardano.Binary (Annotator (..), FullByteString (..), fromCBOR,
-                     serialize, toCBOR)
 import           Cardano.Crypto.DSIGN.Mock (VerKeyDSIGN (..))
 
 import           Ouroboros.Network.Block (BlockNo (..), pattern BlockPoint,
-                     HeaderHash, Point, SlotNo (..), mkSerialised)
+                     Point, SlotNo (..), mkSerialised)
 import           Ouroboros.Network.Point (WithOrigin (..), withOrigin)
 import           Ouroboros.Network.Protocol.LocalStateQuery.Codec (Some (..))
 
 import           Ouroboros.Consensus.Ledger.Abstract
 import           Ouroboros.Consensus.Mempool.API
-import           Ouroboros.Consensus.Storage.ImmutableDB (BinaryInfo (..),
-                     HashInfo (..))
 
 import qualified Shelley.Spec.Ledger.API as SL
 import qualified Shelley.Spec.Ledger.BaseTypes as SL
@@ -74,70 +67,17 @@ import qualified Test.Shelley.Spec.Ledger.NonTraceProperties.Generator as SL
 import           Ouroboros.Consensus.Shelley.Ledger
 import           Ouroboros.Consensus.Shelley.Ledger.History (LedgerViewHistory)
 import qualified Ouroboros.Consensus.Shelley.Ledger.History as History
-import           Ouroboros.Consensus.Shelley.Protocol
 import           Ouroboros.Consensus.Shelley.Protocol.State (TPraosState)
 import qualified Ouroboros.Consensus.Shelley.Protocol.State as TPraosState
 
 import           Generic.Random (genericArbitraryU)
 import           Test.QuickCheck hiding (Result)
 import           Test.QuickCheck.Hedgehog (hedgehog)
-import           Test.Tasty
-import           Test.Tasty.QuickCheck
 
-import           Test.Util.Corruption
 import           Test.Util.Orphans.Arbitrary ()
-import           Test.Util.Roundtrip
 
-import           Test.Consensus.Shelley.Ledger.Golden (mkDummyHash)
-import qualified Test.Consensus.Shelley.Ledger.Golden as Golden
-import           Test.Consensus.Shelley.MockCrypto (TPraosMockCrypto)
-
-tests :: TestTree
-tests = testGroup "Shelley"
-  [ testGroup "Serialisation roundtrips"
-      [ testProperty "roundtrip Block"          prop_roundtrip_Block
-      , testProperty "roundtrip Header"         prop_roundtrip_Header
-      , testProperty "roundtrip HeaderHash"     prop_roundtrip_HeaderHash
-      , testProperty "roundtrip GenTx"          prop_roundtrip_GenTx
-      , testProperty "roundtrip GenTxId"        prop_roundtrip_GenTxId
-      , testProperty "roundtrip ApplyTxErr"     prop_roundtrip_ApplyTxErr
-      , testProperty "roundtrip Query"          prop_roundtrip_Query
-      , testProperty "roundtrip Result"         prop_roundtrip_Result
-      , testProperty "roundtrip ConsensusState" prop_roundtrip_ConsensusState
-      , testProperty "roundtrip LedgerState"    prop_roundtrip_LedgerState
-      ]
-
-  , testProperty "BinaryInfo sanity check" prop_encodeShelleyBlockWithInfo
-
-
-  , testHashInfo (Proxy @TPraosStandardCrypto) "Real crypto"
-  , testHashInfo (Proxy @TPraosMockCrypto)     "Mock crypto"
-
-  , testGroup "Integrity"
-      [ testProperty "generate non-corrupt blocks"  prop_blockIntegrity
-      , testProperty "generate non-corrupt headers" prop_headerIntegrity
-      , testProperty "detect corruption in blocks"  prop_detectCorruption_Block
-      , testProperty "detect corruption in headers" prop_detectCorruption_Header
-      ]
-
-  , Golden.tests
-  ]
-  where
-    testHashInfo :: forall c. Crypto c => Proxy c -> String -> TestTree
-    testHashInfo _ name = testGroup ("HashInfo sanity check: " <> name)
-      [ testProperty "putHash/getHash roundtrip" (prop_shelleyHashInfo_roundtrip @c)
-      , testProperty "hashSize"                  (prop_shelleyHashInfo_hashSize  @c)
-      ]
-
-{-------------------------------------------------------------------------------
-  Block type
--------------------------------------------------------------------------------}
-
--- | We run the tests with mock crypto, as it is easier to generate and debug
--- things. The code is parametric in the crypto, so it shouldn't make much of
--- a difference. This also has the important advantage that we can reuse the
--- generators from cardano-ledger-specs.
-type Block = ShelleyBlock TPraosMockCrypto
+import           Test.Consensus.Shelley.Examples (mkDummyHash)
+import           Test.Consensus.Shelley.MockCrypto
 
 {-------------------------------------------------------------------------------
   SomeResult
@@ -154,132 +94,42 @@ data SomeResult where
 deriving instance Show SomeResult
 
 {-------------------------------------------------------------------------------
-  Serialisation roundtrips
+  Orphans
 -------------------------------------------------------------------------------}
 
-prop_roundtrip_Block :: Block -> Property
-prop_roundtrip_Block = roundtrip' toCBOR ((. Full) . runAnnotator <$> fromCBOR)
+instance Eq (Some (Query Block)) where
+  Some GetLedgerTip == Some GetLedgerTip = True
+  Some GetLedgerTip == _ = False
+  Some GetEpochNo == Some GetEpochNo = True
+  Some GetEpochNo == _ = False
+  Some (GetNonMyopicMemberRewards creds) == Some (GetNonMyopicMemberRewards creds') =
+    creds == creds'
+  Some (GetNonMyopicMemberRewards _) == _ = False
+  Some GetCurrentPParams == Some GetCurrentPParams = True
+  Some GetCurrentPParams == _ = False
+  Some GetProposedPParamsUpdates == Some GetProposedPParamsUpdates = True
+  Some GetProposedPParamsUpdates == _ = False
+  Some GetStakeDistribution == Some GetStakeDistribution = True
+  Some GetStakeDistribution == _ = False
+  Some (GetFilteredUTxO addrs) == Some (GetFilteredUTxO addrs') =
+    addrs == addrs'
+  Some (GetFilteredUTxO _) == _ = False
+  Some GetUTxO == Some GetUTxO = True
+  Some GetUTxO == _            = False
+  Some GetCurrentLedgerState == Some GetCurrentLedgerState = True
+  Some GetCurrentLedgerState == _ = False
+  Some (GetCBOR query) == Some (GetCBOR query') =
+    Some query == Some query'
+  Some (GetCBOR _) == _ = False
 
-prop_roundtrip_Header :: Header Block -> Property
-prop_roundtrip_Header = roundtrip' toCBOR ((. Full) . runAnnotator <$> fromCBOR)
+deriving instance Show (Some (Query Block))
 
-prop_roundtrip_HeaderHash :: HeaderHash Block -> Property
-prop_roundtrip_HeaderHash = roundtrip toCBOR fromCBOR
-
-prop_roundtrip_GenTx :: GenTx Block -> Property
-prop_roundtrip_GenTx = roundtrip toCBOR fromCBOR
-
-prop_roundtrip_GenTxId :: GenTxId Block -> Property
-prop_roundtrip_GenTxId = roundtrip toCBOR fromCBOR
-
-prop_roundtrip_ApplyTxErr :: ApplyTxErr Block -> Property
-prop_roundtrip_ApplyTxErr = roundtrip toCBOR fromCBOR
-
-prop_roundtrip_Query :: Some (Query Block) -> Property
-prop_roundtrip_Query =
-    roundtrip
-      (\case { Some query -> encodeShelleyQuery query })
-      decodeShelleyQuery
-
-prop_roundtrip_Result :: SomeResult -> Property
-prop_roundtrip_Result (SomeResult query result) =
-    roundtrip
-      (encodeShelleyResult query)
-      (decodeShelleyResult query)
-      result
-
-prop_roundtrip_ConsensusState
-  :: TPraosState TPraosMockCrypto
-  -> Property
-prop_roundtrip_ConsensusState = roundtrip toCBOR fromCBOR
-
-prop_roundtrip_LedgerState :: LedgerState Block -> Property
-prop_roundtrip_LedgerState =
-    roundtrip encodeShelleyLedgerState decodeShelleyLedgerState
-
-{-------------------------------------------------------------------------------
-  BinaryInfo
--------------------------------------------------------------------------------}
-
-prop_encodeShelleyBlockWithInfo :: Block -> Property
-prop_encodeShelleyBlockWithInfo blk =
-    encodedHeader === shelleyAddHeaderEnvelope extractedHeader
-  where
-    BinaryInfo { binaryBlob, headerOffset, headerSize } =
-      encodeShelleyBlockWithInfo blk
-
-    extractedHeader :: Lazy.ByteString
-    extractedHeader =
-        Lazy.take (fromIntegral headerSize)   $
-        Lazy.drop (fromIntegral headerOffset) $
-        CBOR.toLazyByteString binaryBlob
-
-    encodedHeader :: Lazy.ByteString
-    encodedHeader = serialize (getHeader blk)
-
-{-------------------------------------------------------------------------------
-  HashInfo
--------------------------------------------------------------------------------}
-
-prop_shelleyHashInfo_roundtrip :: Crypto c => ShelleyHash c -> Property
-prop_shelleyHashInfo_roundtrip h =
-    case Get.runGetOrFail getHash serialisedHash of
-      Left (_, _, e)
-        -> counterexample e $ property False
-      Right (unconsumed, _, h')
-        | Lazy.null unconsumed
-        -> h === h'
-        | otherwise
-        -> counterexample ("unconsumed bytes: " <> show unconsumed) $
-           property False
-  where
-    HashInfo { getHash, putHash } = shelleyHashInfo
-
-    serialisedHash = Put.runPut (putHash h)
-
-prop_shelleyHashInfo_hashSize :: Crypto c => ShelleyHash c -> Property
-prop_shelleyHashInfo_hashSize h =
-    hashSize === fromIntegral (Lazy.length serialisedHash)
-  where
-    HashInfo { hashSize, putHash } = shelleyHashInfo
-
-    serialisedHash = Put.runPut (putHash h)
-
-{-------------------------------------------------------------------------------
-  Integrity
--------------------------------------------------------------------------------}
-
--- TODO test with real crypto
-
-testTPraosSlotsPerKESPeriod :: Word64
-testTPraosSlotsPerKESPeriod = maxBound
-
--- | Test that the block we generate pass the 'verifyBlockIntegrity' check
-prop_blockIntegrity :: Block -> Bool
-prop_blockIntegrity = verifyBlockIntegrity testTPraosSlotsPerKESPeriod
-
--- | Test that the block we generate pass the 'verifyHeaderIntegrity' check
-prop_headerIntegrity :: Header Block -> Bool
-prop_headerIntegrity = verifyHeaderIntegrity testTPraosSlotsPerKESPeriod
-
--- | Test that we can detect random bitflips in blocks.
-prop_detectCorruption_Block :: Block -> Corruption -> Property
-prop_detectCorruption_Block =
-    detectCorruption
-      toCBOR
-      ((. Full) . runAnnotator <$> fromCBOR)
-      (verifyBlockIntegrity testTPraosSlotsPerKESPeriod)
-
--- | Test that we can detect random bitflips in blocks.
-prop_detectCorruption_Header :: Header Block -> Corruption -> Property
-prop_detectCorruption_Header =
-    detectCorruption
-      toCBOR
-      ((. Full) . runAnnotator <$> fromCBOR)
-      (verifyHeaderIntegrity testTPraosSlotsPerKESPeriod)
 
 {-------------------------------------------------------------------------------
   Generators
+
+  These are generators for roundtrip tests, so the generated values are not
+  necessarily valied
 -------------------------------------------------------------------------------}
 
 genHash :: forall a c. Crypto c => Proxy c -> Gen (SL.Hash c a)
@@ -290,7 +140,7 @@ instance Arbitrary Block where
     let Gen.KeySpace_ { Gen.ksCoreNodes } = Gen.geKeySpace Gen.Preset.genEnv
     prevHash         <- unShelleyHash <$> arbitrary
     allPoolKeys      <- elements (map snd ksCoreNodes)
-    txs              <- return [] -- arbitrary
+    txs              <- arbitrary
     curSlotNo        <- SlotNo  <$> choose (0, 10)
     curBlockNo       <- BlockNo <$> choose (0, 100)
     epochNonce       <- arbitrary
@@ -612,34 +462,3 @@ instance Arbitrary (STS.PredicateFailure (STS.LEDGER TPraosMockCrypto)) where
 instance Arbitrary (STS.PredicateFailure (STS.LEDGERS TPraosMockCrypto)) where
   arbitrary = genericArbitraryU
   shrink    = genericShrink
-
-{-------------------------------------------------------------------------------
-  Orphans
--------------------------------------------------------------------------------}
-
-instance Eq (Some (Query Block)) where
-  Some GetLedgerTip == Some GetLedgerTip = True
-  Some GetLedgerTip == _ = False
-  Some GetEpochNo == Some GetEpochNo = True
-  Some GetEpochNo == _ = False
-  Some (GetNonMyopicMemberRewards creds) == Some (GetNonMyopicMemberRewards creds') =
-    creds == creds'
-  Some (GetNonMyopicMemberRewards _) == _ = False
-  Some GetCurrentPParams == Some GetCurrentPParams = True
-  Some GetCurrentPParams == _ = False
-  Some GetProposedPParamsUpdates == Some GetProposedPParamsUpdates = True
-  Some GetProposedPParamsUpdates == _ = False
-  Some GetStakeDistribution == Some GetStakeDistribution = True
-  Some GetStakeDistribution == _ = False
-  Some (GetFilteredUTxO addrs) == Some (GetFilteredUTxO addrs') =
-    addrs == addrs'
-  Some (GetFilteredUTxO _) == _ = False
-  Some GetUTxO == Some GetUTxO = True
-  Some GetUTxO == _            = False
-  Some GetCurrentLedgerState == Some GetCurrentLedgerState = True
-  Some GetCurrentLedgerState == _ = False
-  Some (GetCBOR query) == Some (GetCBOR query') =
-    Some query == Some query'
-  Some (GetCBOR _) == _ = False
-
-deriving instance Show (Some (Query Block))
