+++ ouroboros-consensus-shelley-test/src/Test/ThreadNet/Infra/Shelley.hs
@@ -14,6 +14,8 @@ module Test.ThreadNet.Infra.Shelley (
   , coreNodeKeys
   , mkProtocolRealTPraos
   , tpraosSlotLength
+  , genRealTPraosTestConfig
+  , shrinkRealTPraosTestConfig
   ) where
 
 import           Data.Map.Strict (Map)
@@ -22,6 +24,8 @@ import qualified Data.Sequence.Strict as Seq
 import qualified Data.Set as Set
 import           Data.Time (Day (..), UTCTime (..))
 import           Data.Word (Word64)
+
+import           Test.QuickCheck
 
 import           Cardano.Crypto (ProtocolMagicId (..))
 import           Cardano.Crypto.DSIGN.Class (DSIGNAlgorithm (..), SignKeyDSIGN,
@@ -41,6 +45,11 @@ import           Ouroboros.Consensus.Node.ProtocolInfo
 import           Ouroboros.Consensus.Util.Random
 
 import           Test.Util.Orphans.Arbitrary ()
+
+import           Test.ThreadNet.General
+import           Test.ThreadNet.Util.NodeJoinPlan
+import           Test.ThreadNet.Util.NodeRestarts
+import           Test.ThreadNet.Util.NodeTopology
 
 import qualified Shelley.Spec.Ledger.Address as SL
 import qualified Shelley.Spec.Ledger.BaseTypes as SL
@@ -260,3 +269,27 @@ mkProtocolRealTPraos genesis CoreNode { cnDelegateKey, cnVRF, cnKES, cnOCert } =
         , tpraosIsCoreNodeSignKeyVRF = cnVRF
         }
       }
+
+genRealTPraosTestConfig :: SecurityParam -> Gen TestConfig
+genRealTPraosTestConfig _k = do
+    numCoreNodes <- arbitrary
+    numSlots     <- arbitrary
+
+    -- TODO generate more interesting scenarios
+    let nodeJoinPlan = trivialNodeJoinPlan numCoreNodes
+        nodeTopology = meshNodeTopology numCoreNodes
+
+    initSeed <- arbitrary
+
+    pure TestConfig
+      { nodeJoinPlan
+      , nodeRestarts = noRestarts
+      , nodeTopology
+      , numCoreNodes
+      , numSlots
+      , slotLength = tpraosSlotLength
+      , initSeed
+      }
+
+shrinkRealTPraosTestConfig :: TestConfig -> [TestConfig]
+shrinkRealTPraosTestConfig _ = [] -- TODO
