+++ ouroboros-consensus-shelley-test/test/Main.hs
@@ -2,7 +2,8 @@ module Main (main) where
 
 import           Test.Tasty
 
-import qualified Test.Consensus.Shelley.Ledger (tests)
+import qualified Test.Consensus.Shelley.Golden (tests)
+import qualified Test.Consensus.Shelley.Serialisation (tests)
 import qualified Test.ThreadNet.RealTPraos (tests)
 
 main :: IO ()
@@ -11,6 +12,7 @@ main = defaultMain tests
 tests :: TestTree
 tests =
   testGroup "shelley"
-  [ Test.Consensus.Shelley.Ledger.tests
+  [ Test.Consensus.Shelley.Golden.tests
+  , Test.Consensus.Shelley.Serialisation.tests
   , Test.ThreadNet.RealTPraos.tests
   ]
