+++ ouroboros-consensus-shelley/README.md
@@ -1,5 +1,7 @@
-# Consensus-Shelley
+# ouroboros-consensus-shelley
 
 This package contains:
 
 * `src`: integration of the Shelley ledger with the consensus layer.
+
+For tests, see the `ouroboros-consensus-shelley-test` package.
